Big rework of how the game data is utilised inside the application:

Now generating all spelldata automatically through SimcProfileParser library.
Moved moreData parameter out of all spellservice methods.
Moved conduit data into spells
Updated lookups for spelldata to use spelldata effect values where possible (coeffs, duration, base amounts)
Created test coverage for most of the SpellService methods for individual spells & the base SpellService
Many misc bug fixes
Many new tests introduced

+++ Application/Salvation.Core/Constants/BaseModifier.cs
@@ -3,6 +3,6 @@
     public class BaseModifier
     {
         public string Name;
-        public decimal Value;
+        public double Value;
     }
 }
